CodeChecker: GridGeometry be explicit when using integer division
In this case we are using the behavior of integer division explicitly
and constructing QVector3D with int. Because it accepts floats there is
an implicit conversion, and mixing that with integer division triggers
CodeChecker as a possible error. By explicitly casting to floats we make
it clear this is intentional.

Pick-to: 6.1

+++ src/imports/helpers/gridgeometry.cpp
@@ -243,6 +243,6 @@ void GridGeometry::updateData()
     setStride(32);
     setVertexData(vertexData);
     setPrimitiveType(QQuick3DGeometry::PrimitiveType::Lines);
-    setBounds(QVector3D(-m_vertLines/2, -m_horLines/2, 0.0) * m_horStep,
-                        QVector3D(m_vertLines/2, m_horLines/2, 0.0) * m_vertStep);
+    setBounds(QVector3D(float(-m_vertLines / 2), float(-m_horLines / 2), 0.0f) * m_horStep,
+              QVector3D(float(m_vertLines / 2), float(m_horLines / 2), 0.0f) * m_vertStep);
 }
